workshop/visual/backend/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.32
pydantic>=2.9
python-dotenv>=1.0
orjson>=3.10
pandas>=2.1
pyarrow>=14.0
numpy>=1.24
scipy>=1.11
pyyaml>=6.0
pytest>=8.3
pytest-cov>=5.0
ruff>=0.6
black>=24.10
